refactor: return Result over Option
 * `TryDefault::try_default` now returns `Result<V, DefaultNotFound>`
 * add `DefaultNotFound` error type.

Implements #1
