+++ .gitignore
@@ -1,18 +1,9 @@
-# Generated by Cargo
-# will have compiled files and executables
+# Cargo
 /target/
-
-# Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
-# More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
 
-
-# Added by cargo
-#
-# already existing elements were commented out
-
-/target
-#Cargo.lock
+# MacOS
+.DS_Store
