+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "try_default"
-version = "1.0.1"
+version = "2.0.0"
 authors = ["Joseph Lenton <josephlenton@gmail.com>"]
 edition = "2021"
 license = "MIT"
